packages/ai/src/constants.py
import os

# Load environment variables from .env file
from dotenv import load_dotenv

load_dotenv()

# OpenAI API key
OPENAI_API_KEY = os.getenv('OPENAI_API_KEY')
FALAI_API_KEY = os.getenv('FAL_KEY')
PORT_BACKEND = int(os.getenv('PORT_BACKEND', 3001))

# OpenAI API Headers
HEADERS = {
    "Content-Type": "application/json",
    "Authorization": f"Bearer {OPENAI_API_KEY}"
}