packages/ai/pyproject.toml
[tool.poetry]
name = "ai"
version = "0.1.0"
description = "AI component for Iskay Kimsa Tawa"
authors = ["Your Name <your.email@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
flask = "^2.0.1"
fal-client = "^0.5.0"
python-dotenv = "^1.0.1"
flask-cors = "^5.0.0"
uuid = "^1.30"

[tool.poetry.dev-dependencies]
pytest = "^6.2.5"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"